requirements.txt
Django==5.1.3
django-cors-headers==4.3.1
python-dotenv==1.0.0
supabase==2.10.0
openai==1.12.0
google-generativeai
Pillow
gunicorn
